Renamed all modules and moved some code around for greater consistency.
