+++ tests/test_parallel_layouts.py
@@ -1,12 +1,13 @@
 # Copyright 2025 ACCESS-NRI and contributors. See the top-level COPYRIGHT file for details.
 # SPDX-License-Identifier: Apache-2.0
-"""Tests for access.config.layouts (ComponentLayout and enumeration)."""
+"""Tests for access.config.parallel_layouts (ComponentLayout and enumeration)."""
 
 import dataclasses
 
 import pytest
 
-from access.config.constraints import (
+from access.config.parallel_component import ComponentLayout, ParallelComponent
+from access.config.parallel_constraints import (
     FixedThreadsPerRankConstraint,
     MaxRankFractionConstraint,
     MaxWastedRankFractionConstraint,
@@ -15,12 +16,11 @@ from access.config.constraints import (
     RankRatioGroupConstraint,
     UniformSubdomainConstraint,
 )
-from access.config.domain_parallelisation import Domain, DomainCartesianDecomposition
-from access.config.layouts import (
+from access.config.parallel_domain import Domain, DomainCartesianDecomposition
+from access.config.parallel_layouts import (
+    AllocationStrategy,
     enumerate_layouts,
 )
-from access.config.parallel_component import ComponentLayout, ParallelComponent
-from access.config.parallelisation import AllocationStrategy
 
 # ---------------------------------------------------------------------------
 # Fixtures
@@ -508,3 +508,128 @@ class TestEnumerateLayoutsConstraints:
         for layout in layouts:
             assert layout.decomposition is not None
             assert layout.decomposition.grid[0] % 2 == 0
+
+
+# ---------------------------------------------------------------------------
+# AllocationStrategy
+# ---------------------------------------------------------------------------
+
+
+class TestAllocationStrategy:
+    # --- Fixed mode ---
+
+    def test_fixed_valid(self) -> None:
+        spec = AllocationStrategy(n_ranks=12)
+        assert spec.n_ranks == 12
+        assert spec.allocation_mode == "fixed"
+        assert spec.subcomponents == {}
+
+    def test_fixed_zero_raises(self) -> None:
+        with pytest.raises(ValueError, match=">= 1"):
+            AllocationStrategy(n_ranks=0)
+
+    def test_fixed_negative_raises(self) -> None:
+        with pytest.raises(ValueError, match=">= 1"):
+            AllocationStrategy(n_ranks=-4)
+
+    def test_fixed_with_min_ranks_raises(self) -> None:
+        with pytest.raises(ValueError, match="min_ranks/max_ranks"):
+            AllocationStrategy(n_ranks=4, min_ranks=2)
+
+    def test_fixed_with_max_ranks_raises(self) -> None:
+        with pytest.raises(ValueError, match="min_ranks/max_ranks"):
+            AllocationStrategy(n_ranks=4, max_ranks=8)
+
+    # --- Ratio mode ---
+
+    def test_ratio_valid(self) -> None:
+        spec = AllocationStrategy(weight=3)
+        assert spec.weight == 3
+        assert spec.allocation_mode == "ratio"
+
+    def test_ratio_zero_raises(self) -> None:
+        with pytest.raises(ValueError, match=">= 1"):
+            AllocationStrategy(weight=0)
+
+    def test_ratio_negative_raises(self) -> None:
+        with pytest.raises(ValueError, match=">= 1"):
+            AllocationStrategy(weight=-1)
+
+    def test_ratio_with_min_ranks_raises(self) -> None:
+        with pytest.raises(ValueError, match="min_ranks/max_ranks"):
+            AllocationStrategy(weight=2, min_ranks=2)
+
+    def test_ratio_with_max_ranks_raises(self) -> None:
+        with pytest.raises(ValueError, match="min_ranks/max_ranks"):
+            AllocationStrategy(weight=2, max_ranks=8)
+
+    # --- Both n_ranks and weight raises ---
+
+    def test_fixed_and_ratio_raises(self) -> None:
+        with pytest.raises(ValueError, match="cannot both be set"):
+            AllocationStrategy(n_ranks=4, weight=2)
+
+    # --- Free mode ---
+
+    def test_free_defaults(self) -> None:
+        spec = AllocationStrategy()
+        assert spec.min_ranks == 1
+        assert spec.max_ranks is None
+        assert spec.allocation_mode == "free"
+
+    def test_free_explicit_bounds(self) -> None:
+        spec = AllocationStrategy(min_ranks=4, max_ranks=16)
+        assert spec.min_ranks == 4
+        assert spec.max_ranks == 16
+
+    def test_free_equal_bounds(self) -> None:
+        spec = AllocationStrategy(min_ranks=8, max_ranks=8)
+        assert spec.min_ranks == spec.max_ranks == 8
+
+    def test_free_zero_min_raises(self) -> None:
+        with pytest.raises(ValueError, match=">= 1"):
+            AllocationStrategy(min_ranks=0)
+
+    def test_free_max_less_than_min_raises(self) -> None:
+        with pytest.raises(ValueError, match="max_ranks"):
+            AllocationStrategy(min_ranks=4, max_ranks=2)
+
+    # --- Subcomponents and constraints ---
+
+    def test_with_subcomponents(self) -> None:
+        child_a = AllocationStrategy()
+        child_b = AllocationStrategy(n_ranks=12)
+        parent = AllocationStrategy(subcomponents={"a": child_a, "b": child_b})
+        assert len(parent.subcomponents) == 2
+        assert parent.subcomponents["b"].n_ranks == 12
+
+    def test_with_local_constraints(self) -> None:
+        from access.config.parallel_constraints import FixedThreadsPerRankConstraint
+
+        c = FixedThreadsPerRankConstraint(n_threads=1)
+        spec = AllocationStrategy(local_constraints=(c,))
+        assert spec.local_constraints == (c,)
+
+    def test_group_constraints_default_empty(self) -> None:
+        spec = AllocationStrategy()
+        assert spec.group_constraints == ()
+
+    def test_with_group_constraints(self) -> None:
+        from access.config.parallel_constraints import RankRatioGroupConstraint
+
+        gc = RankRatioGroupConstraint(name_a="a", name_b="b", min_ratio=1.0)
+        spec = AllocationStrategy(group_constraints=(gc,))
+        assert spec.group_constraints == (gc,)
+
+    def test_frozen(self) -> None:
+        import dataclasses
+
+        spec = AllocationStrategy()
+        with pytest.raises(dataclasses.FrozenInstanceError):
+            spec.n_ranks = 4  # type: ignore[misc]
+
+    def test_hash_includes_structure(self) -> None:
+        child = AllocationStrategy(n_ranks=2)
+        spec_a = AllocationStrategy(subcomponents={"child": child})
+        spec_b = AllocationStrategy(subcomponents={"child": child})
+        assert hash(spec_a) == hash(spec_b)
